80 perc rewritten to better cpp code
