+++ include/socket.hpp
@@ -11,22 +11,23 @@
 class Socket
 {
 	private:
-		//struct	sockaddr_in	servAddr;
-		int	listenfd;
-		int kq;
-		struct kevent evSet;
 		unsigned int port;
-		struct uc {
-			int uc_fd;
-			char *uc_addr;
-		} users[10]; //10 is hardcoded!!
-	public:
-		Socket(unsigned short newport);
-		bool	setUpConn();
+		// struct uc {
+		// 	int uc_fd;
+		// 	char *uc_addr;
+		// } users[10]; //10 is hardcoded!!
+		// ^^ ?? if I use this instead of global, it fails!? ^^
+		struct kevent evSet;
+		int kq;
+		struct	sockaddr_in	servAddr;
+		int	listenfd;
 		void	watchLoop();
 		int		connIndex(int fd);
 		int		connAdd(int fd);
 		int		connDelete(int fd);
+	public:
+		Socket(unsigned short newport);
+		bool	setUpConn();
 };
 
 #endif
