+++ src/server_old.cpp
@@ -1,5 +1,4 @@
 #include "../include/server.hpp"
-#include "../include/socket.hpp"
 #include <cstdio>
 #include <cstring>
 #include <arpa/inet.h>
@@ -8,12 +7,12 @@
 #include <sys/event.h>
 #include <fstream>
 
-// #define NUSERS 10
+#define NUSERS 10
 
-// struct uc { //this should go in the class!!
-//     int uc_fd;
-//     char *uc_addr;
-// } users[NUSERS];
+struct uc { //this should go in the class!!
+    int uc_fd;
+    char *uc_addr;
+} users[NUSERS];
 
 
 bool	write_exit(std::string error)
@@ -23,23 +22,22 @@ bool	write_exit(std::string error)
 }
 
 /* find the index of a file descriptor or a new slot if fd=0 */
-int	Socket::connIndex(int fd) 
+int	conn_index(int fd) 
 {
     int uidx;
-    for (uidx = 0; uidx < 10; uidx++)
+    for (uidx = 0; uidx < NUSERS; uidx++)
         if (users[uidx].uc_fd == fd)
             return uidx;
     return -1;
 }
 
-int Socket::connAdd(int fd) 
+int conn_add(int fd) 
 {
     int uidx;
-    if (fd < 1) 
-		return -1;
-    if ((uidx = this->connIndex(0)) == -1)
+    if (fd < 1) return -1;
+    if ((uidx = conn_index(0)) == -1)
         return -1;
-    if (uidx == 10) {
+    if (uidx == NUSERS) {
         close(fd);
         return -1;
     }
@@ -49,11 +47,11 @@ int Socket::connAdd(int fd)
 }
 
 /* remove a connection and close it's fd */
-int Socket::connDelete(int fd) {
+int conn_del(int fd) {
     int uidx;
     if (fd < 1) 
 		return -1;
-    if ((uidx = connIndex(fd)) == -1)
+    if ((uidx = conn_index(fd)) == -1)
         return -1;
 
     users[uidx].uc_fd = 0;
@@ -63,7 +61,7 @@ int Socket::connDelete(int fd) {
     return close(fd);
 }
 
-void	Socket::watchLoop()
+void	watch_loop(int kq, int listenfd)
 {
 	struct kevent evSet;
 	struct kevent evList[32];
@@ -71,19 +69,9 @@ void	Socket::watchLoop()
 	struct sockaddr_storage addr;
 	socklen_t socklen = sizeof(addr);
 	int fd;
-	this->kq = kqueue();
-	printf("kq = %d\n", kq);
-	EV_SET(&evSet, listenfd, EVFILT_READ, EV_ADD, 0, 0, NULL);//EV_SET is a macro that fills the kevent struct
-	if (kevent(this->kq, &evSet, 1, NULL, 0, NULL) == -1)
-		return ;//(write_exit("kqueue/kevent error"));
 	while (1)
 	{
-		printf("TEST\n");
-		printf("kq there = %d\n", kq);
-
-		nev = kevent(this->kq, NULL, 0, evList, 32, NULL);
-		printf("TEST2\n");
-
+		nev = kevent(kq, NULL, 0, evList, 32, NULL);
 		if (nev < 1)
 		{
 			write_exit("kevent error");
@@ -101,10 +89,9 @@ void	Socket::watchLoop()
 					write_exit("kevent error");
 					return ;
 				}
-				// int uidx;
 				//users[]
+				conn_del(fd);
 				// close(fd);
-				connAdd(fd);
 			}
 			else if ((int)evList[i].ident == listenfd)
 			{
@@ -116,7 +103,7 @@ void	Socket::watchLoop()
 					write_exit("accept error");
 					return ;
 				}
-				if (this->connAdd(fd) == 0)
+				if (conn_add(fd) == 0)
 				{
 					EV_SET(&evSet, fd, EVFILT_READ, EV_ADD, 0, 0, NULL);
 					if (kevent(kq, &evSet, 1, NULL, 0, NULL) == -1)
@@ -155,65 +142,36 @@ void	Socket::watchLoop()
 	}
 }
 
-bool	Socket::setUpConn()
+bool	Server::start()
 {
+	if (this->running)
+		return (false);
+	int		listenfd;//.n;//connfd; //connfd will actually talk to the client that's connected
 	struct	sockaddr_in	servAddr;
 
 	if ((listenfd = socket(AF_INET, SOCK_STREAM, 0)) < 0) //AF_INET = internet socket, SOCK_STREAM = tcp stream
 		return (write_exit("socket error"));
+	int reuse; //this and setsockopt avoids the bind error and allows to reuse the address
+	if (setsockopt(listenfd, SOL_SOCKET, SO_REUSEADDR, (char *)&reuse, sizeof(int)) == -1)
+ 		return(write_exit("reuse port error"));
 	//setting up address you're listening on
 	std::memset(&servAddr, '\0', sizeof(servAddr));
 	servAddr.sin_family		= AF_INET;
 	servAddr.sin_addr.s_addr = htonl(INADDR_ANY); // will respond to anything
-	servAddr.sin_port		= htons(port); //port you're listening on
+	servAddr.sin_port		= htons(SERVER_PORT); //port you're listening on
 
 	if ((bind(listenfd, (SA *) &servAddr, sizeof(servAddr))) < 0)//bind listening socket to address
 		return (write_exit("bind error"));
 	if ((listen(listenfd, 10)) < 0)
 		return (write_exit("listen error"));
-	// this->kq = kqueue();
-	// printf("kq = %d\n", kq);
-	// EV_SET(&evSet, listenfd, EVFILT_READ, EV_ADD, 0, 0, NULL);//EV_SET is a macro that fills the kevent struct
-	// if (kevent(this->kq, &evSet, 1, NULL, 0, NULL) == -1)
-	// 	return (write_exit("kqueue/kevent error"));
-	this->watchLoop();
-	return (true);
-}
 
-bool	Server::start()
-{
-	if (this->running)
-		return (false);
-	unsigned short testprt = 80;
-	Socket	sck(testprt);
-	sck.setUpConn();
-
-	//std::for_each(sockets.cbegin(), sockets.cend(), Socket::setUpConn());
-	
-
-	// int		listenfd;//.n;//connfd; //connfd will actually talk to the client that's connected
-	// struct	sockaddr_in	servAddr;
-
-	// if ((listenfd = socket(AF_INET, SOCK_STREAM, 0)) < 0) //AF_INET = internet socket, SOCK_STREAM = tcp stream
-	// 	return (write_exit("socket error"));
-	// //setting up address you're listening on
-	// std::memset(&servAddr, '\0', sizeof(servAddr));
-	// servAddr.sin_family		= AF_INET;
-	// servAddr.sin_addr.s_addr = htonl(INADDR_ANY); // will respond to anything
-	// servAddr.sin_port		= htons(SERVER_PORT); //port you're listening on
-
-	// if ((bind(listenfd, (SA *) &servAddr, sizeof(servAddr))) < 0)//bind listening socket to address
-	// 	return (write_exit("bind error"));
-	// if ((listen(listenfd, 10)) < 0)
-	// 	return (write_exit("listen error"));
-
-	// int kq;
-	// struct kevent evSet;
-	// kq = kqueue();
-	// EV_SET(&evSet, listenfd, EVFILT_READ, EV_ADD, 0, 0, NULL);//EV_SET is a macro that fills the kevent struct
-	// if (kevent(kq, &evSet, 1, NULL, 0, NULL) == -1)
-	// 	return (write_exit("kqueue/kevent error"));
-	// watch_loop(kq, listenfd);
+	int kq;
+	struct kevent evSet;
+	kq = kqueue();
+	EV_SET(&evSet, listenfd, EVFILT_READ, EV_ADD, 0, 0, NULL);//EV_SET is a macro that fills the kevent struct
+	if (kevent(kq, &evSet, 1, NULL, 0, NULL) == -1)
+		return (write_exit("kqueue/kevent error"));
+	watch_loop(kq, listenfd);
 	return (true);	
 }
 
@@ -224,4 +182,3 @@ bool	Server::start()
 // 	server.start(ipAddress, port);
 // 	server.stop();
 // }
-
